move remove button into own component
